+++ lib/utils/addon-utils.ts
@@ -1,0 +1,11 @@
+
+import { ClusterAddOn } from '../spi';
+
+/**
+ * Returns AddOn Id if defined else returns the class name
+ * @param addOn
+ * @returns string
+ */
+export function getAddOnNameOrId(addOn: ClusterAddOn): string {
+  return addOn.id ?? addOn.constructor.name;
+}
